+++ Tests/Runtime/MeshOps/CollapseVerticesTests.cs
@@ -3,7 +3,10 @@ using UnityEngine;
 using UnityEngine.ProBuilder;
 using UnityEngine.ProBuilder.MeshOperations;
 using UnityEngine.ProBuilder.Shapes;
+using UnityEngine.ProBuilder.Tests;
+#if UNITY_EDITOR && PB_CREATE_TEST_MESH_TEMPLATES
 using UnityEngine.ProBuilder.Tests.Framework;
+#endif
 
 static class CollapseVerticesTests
 {
@@ -18,14 +21,13 @@ static class CollapseVerticesTests
         cube.ToMesh();
         cube.Refresh();
 
-#if PB_CREATE_TEST_MESH_TEMPLATES
+#if UNITY_EDITOR && PB_CREATE_TEST_MESH_TEMPLATES
         TestUtility.SaveAssetTemplate(cube.mesh, cube.name);
 #endif
-
-        TestUtility.AssertMeshAttributesValid(cube.mesh);
-        var template = TestUtility.GetAssetTemplate<Mesh>(cube.name);
+        RuntimeUtility.AssertMeshAttributesValid(cube.mesh);
+        var template = Resources.Load<Mesh>(RuntimeUtility.GetResourcesPath<Mesh>(cube.name));
         Assert.IsNotNull(template);
-        TestUtility.AssertAreEqual(template, cube.mesh);
+        RuntimeUtility.AssertAreEqual(template, cube.mesh);
 
         Object.DestroyImmediate(cube);
     }
@@ -39,14 +41,13 @@ static class CollapseVerticesTests
         cube.ToMesh();
         cube.Refresh();
 
-#if PB_CREATE_TEST_MESH_TEMPLATES
+#if UNITY_EDITOR && PB_CREATE_TEST_MESH_TEMPLATES
         TestUtility.SaveAssetTemplate(cube.mesh, cube.name);
 #endif
-
-        TestUtility.AssertMeshAttributesValid(cube.mesh);
-        var template = TestUtility.GetAssetTemplate<Mesh>(cube.name);
+        RuntimeUtility.AssertMeshAttributesValid(cube.mesh);
+        var template = Resources.Load<Mesh>(RuntimeUtility.GetResourcesPath<Mesh>(cube.name));
         Assert.IsNotNull(template);
-        TestUtility.AssertAreEqual(template, cube.mesh);
+        RuntimeUtility.AssertAreEqual(template, cube.mesh);
 
         UnityEngine.Object.DestroyImmediate(cube);
     }